deployment/requirements.txt
uvicorn
fastapi
beautifulsoup4
lxml
requests
tensorflow
python-dotenv
firebase_admin
numpy
pandas
python-multipart
langchain_community
langchain
boto3
chromadb
langchain_aws
langchain_chroma
python-dotenv
langchain_core
huggingface_hub
langchain_google_genai
langchain_openai
langchain_anthropic
google-generativeai
openai